Removed Forest.hdr Removed invert colours as it is no longer needed. Added an interface to allow items to be drawn to the ui. Post-processing is now part of the camera rather than the rendering pipeline. Mass refactoring Changed the renderer to be a class. Added colour grading shader. Exposed the post-processing settings to the user. Added bloom from my previous project. The start of the post-processing stack. Changed the contents of the scene. Added specular IBL. Some minor refactorings. Added new hdr texture to better follow along with the tutorial. Added roughness to the ambient equation. Fixed samplers not being binded correctly. Added irradiance mapping to ambient lighting. Added support for HDR files to be converted into cube maps. Fixed ambient not being muted down. Removed the specular texture buffer as it is no longer used. Added BRDF shading to directional lights. Added sRGB gamma correction. Added parallax occlusion mapping. Moved some code from the fragment shader to the vertex shader. Fixed view_direction.y being the wrong way round. Added parallax mapping.

+++ include/engine/loader/RawMesh.h
@@ -107,7 +107,7 @@ namespace load
         for (const ObjVertex &vertex : intermediateVertices)
         {
             mVertices.emplace_back(vertex);  // implicit conversion to user defined type.
-            uniqueIndices.emplace_back(mVertices.size() - 1);
+            uniqueIndices.emplace_back(static_cast<uint32_t>((mVertices.size() - 1)));
         }
        
         createIndices(uniqueIndices);
